+++ src/tasks/windows.rs
@@ -5,7 +5,7 @@ use utils;
 pub fn sync() {
     println!("pkg: windows: syncing ...");
 
-    let bin = utils::env::home_dir().join(Path::new("bin"));
+    let bin = utils::env::home_dir().join(Path::new(".local/bin"));
     utils::fs::create_dir_all_or_panic(Some(&bin));
     println!("{:?}", &bin);